Add suport for polygon as extent and reorganise tests (#48)
* Add suport for polygon as extent and reorganise tests

* Update 1_3_2_bro_gpd.ipynb

* Add coverage and coverage buton in readme

* Update pyproject.toml
