+++ .github/workflows/ci.yml
@@ -13,6 +13,15 @@ on:
       - main
       - dev
 
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+concurrency:
+  group: pages
+  cancel-in-progress: false
+
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -35,4 +44,30 @@ jobs:
 
       - name: Run tests
         run: |
-          pytest ./tests
+          pytest ./tests --cov=brodata --cov-report=xml --cov-report=html --cov-report=term-missing
+
+      - name: Upload coverage preview artifact (dev)
+        if: github.event_name == 'push' && github.ref == 'refs/heads/dev'
+        uses: actions/upload-artifact@v4
+        with:
+          name: coverage-html-dev
+          path: htmlcov
+          retention-days: 7
+
+      - name: Upload coverage artifact
+        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+        uses: actions/upload-pages-artifact@v4
+        with:
+          path: htmlcov
+
+  deploy-coverage:
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    runs-on: ubuntu-latest
+    needs: build
+    steps:
+      - name: Deploy coverage report to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@v4
